+++ gspa-predictors/src/test/groovy/gspa/predictor/neural/NeuralSidecarPredictorsSpec.groovy
@@ -168,6 +168,73 @@ p1\tEC:1.1.1.1\t0.65\tEC
         results['p1'].any { it.type == AnnotationType.EC && it.value == 'EC:1.1.1.1' }
     }
 
+    def "CafaBaselinePredictor: buildCommand serialises integrator + components-dir + dag"() {
+        given:
+        def predictor = new CafaBaselinePredictor(
+            sidecarScript: writeSidecarStub().absolutePath,
+            integrator: '/tmp/integrator.json',
+            componentsDir: '/tmp/components',
+            dag: '/tmp/go-dag.tsv',
+        )
+        def fasta = tmp.resolve('q.faa').toFile()
+        fasta.text = '>p1\nMKTAY\n'
+        def outDir = tmp.resolve('outcb').toFile()
+        outDir.mkdirs()
+
+        when:
+        def cmd = predictor.buildCommand(fasta, outDir)
+
+        then:
+        cmd[cmd.indexOf('--predictor') + 1] == 'cafa-baseline'
+        cmd.contains('--integrator')
+        cmd[cmd.indexOf('--integrator') + 1] == '/tmp/integrator.json'
+        cmd.contains('--components-dir')
+        cmd[cmd.indexOf('--components-dir') + 1] == '/tmp/components'
+        cmd.contains('--dag')
+        cmd[cmd.indexOf('--dag') + 1] == '/tmp/go-dag.tsv'
+        predictor.outputTypes == ([AnnotationType.GO] as Set)
+    }
+
+    def "CafaBaselinePredictor: parseOutput reads the 4-column TSV"() {
+        given:
+        def predictor = new CafaBaselinePredictor(
+            sidecarScript: writeSidecarStub().absolutePath,
+            integrator: '/tmp/i.json', componentsDir: '/tmp/c', dag: '/tmp/d',
+        )
+        def outDir = tmp.toFile()
+        writeOutputTsv('query.cafa-baseline.tsv', '''\
+protein_id\tterm\tscore\tannotation_type
+p1\tGO:0003824\t0.512\tGO
+p1\tGO:0008150\t0.331\tGO
+'''.stripIndent())
+
+        when:
+        def results = predictor.parseOutput(outDir)
+
+        then:
+        results['p1'].size() == 2
+        results['p1'][0].type == AnnotationType.GO
+        results['p1'][0].value == 'GO:0003824'
+        results['p1'][0].score == 0.512
+        results['p1'][0].source == 'cafa-baseline'
+    }
+
+    def "CafaBaselinePredictor: buildCommand fails fast without integrator/components/dag"() {
+        given:
+        def predictor = new CafaBaselinePredictor(
+            sidecarScript: writeSidecarStub().absolutePath,
+        )  // missing integrator + componentsDir + dag
+        def fasta = tmp.resolve('q.faa').toFile()
+        fasta.text = '>p1\nM\n'
+        def outDir = tmp.toFile()
+
+        when:
+        predictor.buildCommand(fasta, outDir)
+
+        then:
+        thrown(IllegalStateException)
+    }
+
     def "buildCommand fails fast when required config is missing"() {
         given:
         def sidecarOk = new DeepGoPlusEsm2Predictor(